requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
dj-static==0.0.6
Django==2.0.4
django-cors-headers==2.2.0
django-material==1.2.3
djangorestframework==3.8.2
future==0.16.0
gunicorn==19.7.1
idna==2.6
oauthlib==2.0.7
pycryptodomex==3.6.1
pyjwkest==1.4.0
PyJWT==1.6.1
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
virtualenv==15.1.0